Introduce delegates with shader modification
This is a large change that introduces the delegate concept. Each
pipeline can now define a delegate type that must be provided when the
pipeline is made. The delegate contains members that can be assigned to
expressions using the `delegate!` macro. The expressions can contain
references to buffers. The pipelines in turn modify their shaders to
reflect these new expressions and bindings. This allows for easy
modification of the visualization.