Cleanup derive and simplify API usage
This change keeps reference to buffers around so the user does not need to pass them in during the
render stage.
Hopefully I will be able to make this API a bit less implicit later,
but for now I think using a Rc<RefCell<BufferInner>> makes the API way easier to use and is worth
the loss of explicitness.