Remove delegate macro in favor of operator overloading
This change introduces value types that are chained with operator
overloading to provide the necessary features. This removes the need for
the delegate! macro. Some features, such as vectors, are still a bit
cumbersome to implement. Further, this does not yet provide compile-time
type-checking. This might be introduced in a later change.