Improve speed in molecular_dynamics example
With this change we ensure that we do not trigger too many updates on
systems with high update rates (no frame rate limit).
Instead, we lock the update frequency to a target frame rate.
Also ensure that the time since last update cannot grow exponentially
larger if updates are slower than expected.